Everything works

+++ src/java_project/Vehicle.java
@@ -2,7 +2,7 @@ package java_project;
 public class Vehicle 
 {
    protected String vID;//Μοναδικό ID του οχήματος
-   protected int vPrice;//τιμή οχήματος ανα μέρα
+   protected Double vPrice;//τιμή οχήματος ανα μέρα
    protected int vWheels;//αριθμός τροχών οχήματος
    
    public Vehicle(String vID)
@@ -15,7 +15,7 @@ public class Vehicle
            return this.vID;
    }
    
-   public int getPrice()
+   public Double getPrice()
    {//επιστρέφει την τιμή
        return this.vPrice;
    }
